+++ src/analytics.ts
@@ -14,7 +14,8 @@ enum Category {
 	Extension,
 	TODOs,
 	Analyzer,
-	Debugger
+	Debugger,
+	LintNames
 }
 
 enum EventAction {
@@ -39,6 +40,7 @@ class Analytics {
 	};
 	logSdkDetectionFailure() { this.event(Category.Extension, EventAction.SdkDetectionFailure); }
 	logShowTodosToggled(enabled: boolean) { this.event(Category.TODOs, enabled ? EventAction.Enabled : EventAction.Disabled); }
+	logLintNamesToggled(enabled: boolean) { this.event(Category.LintNames, enabled ? EventAction.Enabled : EventAction.Disabled); }
 	logAnalyzerError(description: string, fatal: boolean) { this.error("AS: " + description, fatal); }
 	logAnalyzerStartupTime(timeInMS: number) { this.time(Category.Analyzer, TimingVariable.Startup, timeInMS); }
 	logAnalyzerFirstAnalysisTime(timeInMS: number) { this.time(Category.Analyzer, TimingVariable.FirstAnalysis, timeInMS); }