+++ src/commands/edit.ts
@@ -16,17 +16,28 @@ export class EditCommands implements vs.Disposable {
 
 		this.commands.push(
 			vs.commands.registerTextEditorCommand("dart.organizeDirectives", this.organizeDirectives, this),
-			vs.commands.registerCommand("dart.applySourceChange", this.applyEdits, this)
+			vs.commands.registerTextEditorCommand("dart.sortMembers", this.sortMembers, this),
+			vs.commands.registerCommand("_dart.applySourceChange", this.applyEdits, this)
 		);
 	}
 
 	private organizeDirectives(editor: vs.TextEditor, editBuilder: vs.TextEditorEdit) {
+		this.sendEdit(this.analyzer.editOrganizeDirectives, "Organize Directives", editor, editBuilder);
+	}
+
+	private sortMembers(editor: vs.TextEditor, editBuilder: vs.TextEditorEdit) {
+		this.sendEdit(this.analyzer.editSortMembers, "Sort Members", editor, editBuilder);
+	}
+
+	private sendEdit(f: (a: { file: string }) => Thenable<{edit: as.SourceFileEdit}>, commandName: string, editor: vs.TextEditor, editBuilder: vs.TextEditorEdit) {
 		if (!editors.hasActiveDartEditor()) {
 			vs.window.showWarningMessage("No active Dart editor.");
 			return;
 		}
 
-		this.analyzer.editOrganizeDirectives({ file: editor.document.fileName }).then((response) => {
+		f = f.bind(this.analyzer); // Yay JavaScript!
+
+		f({ file: editor.document.fileName }).then((response) => {
 			let edit: as.SourceFileEdit = response.edit;
 			if (edit.edits.length == 0)
 				return;
@@ -41,10 +52,10 @@ export class EditCommands implements vs.Disposable {
 				});
 			}).then((result) => {
 				if (!result)
-					vs.window.showWarningMessage("Unable to apply organize directives edits.");
+					vs.window.showWarningMessage(`Unable to apply ${commandName} edits.`);
 			});
 		}, (error) => {
-			vs.window.showErrorMessage(`Error running organize directives: ${error.message}.`);
+			vs.window.showErrorMessage(`Error running ${commandName}: ${error.message}.`);
 		});
 	}