+++ src/debug/flutter_run.ts
@@ -59,12 +59,16 @@ export class FlutterRun extends StdIOService {
 
 	// Request methods.
 
-	restart(appId: string, pause: boolean): Thenable<any> {
-		return this.sendRequest("app.restart", { "appId": appId, fullRestart: false, pause: pause });
+	restart(appId: string, pause: boolean, fullRestart?: boolean): Thenable<any> {
+		return this.sendRequest("app.restart", { "appId": appId, fullRestart: fullRestart === true, pause: pause });
 	}
 
 	stop(appId: string): Thenable<UnknownResponse> {
 		return this.sendRequest("app.stop", { "appId": appId });
+	}
+
+	callServiceExtension(appId: string, methodName: string, params: any): Thenable<any> {
+		return this.sendRequest("app.callServiceExtension", { "appId": appId, methodName: methodName, params: params });
 	}